+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2022.05.09',
+  '2022.05.12',
   // Fallback runtime versions.
+  '2022.05.09',
   '2022.05.03',
-  '2022.04.21',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -53,11 +53,11 @@ bool shouldGCVersion(String version) =>
     version.compareTo(gcBeforeRuntimeVersion) < 0;
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
-final String runtimeSdkVersion = '2.16.1';
+final String runtimeSdkVersion = '2.17.0';
 final String toolStableDartSdkVersion = '2.17.0';
-final String toolStableFlutterSdkVersion = '2.10.5';
+final String toolStableFlutterSdkVersion = '3.0.0';
 final String toolPreviewDartSdkVersion = '2.17.0';
-final String toolPreviewFlutterSdkVersion = '2.13.0-0.4.pre';
+final String toolPreviewFlutterSdkVersion = '3.0.0';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);